add email config save/restore
new registrations and validations call save when the map is
updated. during initialization the save file is read.

to be sure we don't accidently write garbage to the save file, new
file writes are saved to a temporary file. the temp file is renamed to
the expected save file name when complete.