reserve - add end command, fix modified time handling
round last modified value for better comparison since the RFC1123
format trims subsecond accuracy.

added last modified check to delete in server

in delete and end commands, show loans correctly

new end command is intended to close/end an active reservation. active
means it's on loan or currently in progress. a lookup tries to find
the resource then issues a delete API request to end it. the delete
API is used for all reservation termination and deletion of future
events.